+++ src/civil/Civil.java
@@ -1,6 +1,5 @@
 package civil;
 
-import payment.Discount;
 import ticket.Category;
 import ticket.GovAuthorities;
 import ticket.Ticket;
