+++ src/commercial/Commercial.java
@@ -1,6 +1,5 @@
 package commercial;
 
-import payment.Discount;
 import ticket.Category;
 import ticket.GovAuthorities;
 import ticket.Ticket;
