+++ src/payment/Charges.java
@@ -1,16 +1,16 @@
 package payment;
 
-import ticket.DatesAssignment;
+import datetimeutils.DateTimeUtils;
 
 import java.time.temporal.ChronoUnit;
 
 public class Charges {
 
-    public static long daysLate = ChronoUnit.DAYS.between(Discount.paymentDeadline, DatesAssignment.currentDate);
+    public static long daysLate = ChronoUnit.DAYS.between(ApplyDiscount.paymentDeadline, DateTimeUtils.currentDate);
 
 
     public static double calculateInterest(double fineAmount, long daysLate) {
-        if (daysLate <= Discount.DEADLINE) {
+        if (daysLate <= ApplyDiscount.DEADLINE) {
         double interestRate = 0.05;
         return fineAmount * Math.pow(1 + interestRate, daysLate);
         }
